1.1 Inform User of Error detection
1.1 Brings enhanced error detection, informing the user if they have
entered a character which is unsupported in Morse Code (all
alpha-numeric characters are supported, though many special
characters/symbols are not). 1.0 would just skip any character it didn't
recognize (since the timing array is all 0s or doesn't exist) without
informing the user.

1.1 Also starts an attemp to show the user what is being displayed.
However, there is a bug in the code that causes the promp for user input
to display several times before it is needed again. The line breaks
caused by this interrupt any chance to tell the user what is being
displayed. Working on a fix!