1.1.5 Logic bug fix + small note change
In addition to some note changes, there was a potential logic bug in the
code where the code might fail to present the text in Morse if only a
single character was entered. This resulted from the way I initialized
my linked-list (I did not have a root node). I changed the
implementation of the linked-list to include a root node to solve this
problem.