1.2 Major bug fixes
Fixed two major bugs.

1. The first bug that was fixed was the bug where the program would ask
for additional user input while still processing the previous user
input, causing the display flow to be incorrect. This was fixed by
adding a small delay(100) so that the program has time to read in every
byte entered by the user.

2. There was a logic bug that caused lower case letters to be counted as
characters that were not in the Morse "dictionary." I moved the
conversion of lower case letters to upper case letters before the "in
dictionary" check.